requirements.txt
allure-pytest==2.8.36
allure-python-commons==2.8.36
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
importlib-metadata==3.7.3
importlib-resources==5.1.2
iniconfig==1.1.1
jdcal==1.4.1
jsonpath==0.82
openpyxl==2.6.2
packaging==20.9
pipenv==2020.11.15
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==1.9.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
